refactor hierarchy module
